Use programs from the examples folder for tests where possible
