+++ examples/src/main/java/com/aws/verifier/examples/BinarySearch.java
@@ -1,52 +1,47 @@
 package com.aws.verifier.examples;
 
-import com.aws.jverify.*;
-
 import static com.aws.jverify.JVerify.*;
 
+import com.aws.jverify.Erased;
+import com.aws.jverify.Pure;
+
 class BinarySearch {
-  @Erased 
-  @Pure
-  boolean sorted(int[] arr)
-  {
-    reads(arr);
+    @Pure
+    @Erased
+    static boolean sorted(int[] arr) {
+        reads(arr);
+        return forall((Integer i, Integer j) ->
+                !(0 <= i && i < j && j < arr.length) || arr[i] < arr[j]);
+    }
 
-    return forall((Integer i) -> forall((Integer j) -> 
-      implies(0 <= i && i < j && j < arr.length, arr[i] < arr[j])));
-  }
+    public static int binarySearchImpl(int[] arr, int key) {
+        precondition(arr.length <= 0x7fff_ffff /* Integer.MAX_VALUE */);
+        precondition(sorted(arr));
+        postcondition((Integer res) ->
+                (res == -1 && !sequence(arr).contains(key))
+                || (0 <= res && res < arr.length && arr[res] == key)
+        );
 
-  int binarySearch(int[] arr, int target) {
-      precondition(sorted(arr));
-      postcondition((Integer r) -> sequence(arr).contains(target) 
-        ? 0 <= r && r < arr.length && arr[r] == target 
-        : r == -1
-      );
+        var lo = 0;
+        var hi = arr.length;
 
-      var left = 0;
-      var right = arr.length - 1; // Bug: should be arr.Length
-      
-      while (left < right) 
-      {
-        invariant(0 <= left && left <= right && right <= arr.length);
-        invariant(!sequence(arr).subsequence(left, right).contains(target));
+        while (lo < hi) {
+            invariant(0 <= lo);
+            invariant(lo <= hi);
+            invariant(hi <= arr.length);
+            invariant(!sequence(arr).take(lo).contains(key));
+            invariant(!sequence(arr).drop(hi).contains(key));
 
-        var mid = (left + right) / 2;
-        if (arr[mid] == target) {
-//          ^^^^^^^^ JSpec error: index out of range
-            return mid;
+            var mid = lo + (hi - lo) / 2;
+            if (key < arr[mid]) {
+                hi = mid;
+            } else if (arr[mid] < key) {
+                lo = mid + 1;
+            } else {
+                return mid;
+            }
         }
-        if (arr[mid] < target) {
-            left = mid + 1;
-        } else {
-            right = mid - 1; // Bug: should be just mid
-        }
-      }
-      return -1;
-  }
 
-  @Pure
-  boolean implies(boolean ante, boolean cons) {
-    return !ante || cons;
-  }
-
+        return -1;
+    }
 }
