+++ examples/src/main/java/com/aws/verifier/examples/Fibonacci.java
@@ -1,27 +1,27 @@
 package com.aws.verifier.examples;
 
-import com.aws.jverify.*;
 import static com.aws.jverify.JVerify.*;
 
-@SuppressWarnings("ConstantValue")
+import com.aws.jverify.Erased;
+import com.aws.jverify.Nat;
+import com.aws.jverify.Proof;
+import com.aws.jverify.Pure;
+import com.aws.jverify.Unbounded;
+
 class Fibonacci {
-    @Pure // Enables Spec to be called in contract calls
-    @Erased // Prevents Spec from being called at runtime, and allows use of @Unbounded
-    // @Unbounded specifies that the int will not overflow
-    // @Nat specified that the int is non-negative
+    @Pure
+    @Erased
     static @Unbounded @Nat int Spec(@Unbounded @Nat int n) {
         return n < 2 ? n : Spec(n - 1) + Spec(n - 2);
     }
 
     public static @Nat int Implementation(@Nat int n)
     {
-        precondition(Spec(n) <= 0x7fffffff/*Integer.MAX_VALUE*/);
+        precondition(Spec(n) <= 0x7fffffff /*Integer.MAX_VALUE*/);
         postcondition((Integer r) -> r == Spec(n));
-//      ^ Related location: this is the postcondition that could not be proved
 
         if (n == 0) {
-            return 1;
-//                 ^ Error: a postcondition could not be proved on this return path
+            return 0;
         }
 
         int previousResult = 0;
@@ -48,9 +48,7 @@ class Fibonacci {
         precondition(i <= j);
         postcondition(Spec(i) <= Spec(j));
 
-        @Unbounded 
-        @Nat 
-        int x = i;
+        @Unbounded @Nat int x = i;
         while(x < j)
         {
             invariant(x <= j);
