+++ verifier/src/main/java/com/aws/jverify/verifier/Driver.java
@@ -58,7 +58,9 @@ public class Driver {
                 // --show-snippets has no affect because Dafny can't extract them from the serialized source anyways
                 processBuilder.command().add("--show-snippets=false");
                 processBuilder.command().add("--print-ranges");
-            }                
+            }
+            processBuilder.command().add("--ignore-indentation");
+            
             for(var option : verifierOptions.additionalDafnyArguments()) {
                 processBuilder.command().add(option);
             }
